# Conflicts:
#	Engine/AbortableRenderInfo.cpp
#	Engine/AbortableRenderInfo.h
#	Engine/BlockingBackgroundRender.cpp
#	Engine/CacheEntry.h
#	Engine/FrameEntry.cpp
#	Engine/FrameKey.cpp
#	Engine/FrameKey.h
#	Engine/ImageKey.cpp
#	Engine/ImageLocker.h
#	Engine/ImageParams.h
#	Engine/KeyHelper.h
#	Engine/Node.cpp
#	Engine/OutputEffectInstance.cpp
#	Engine/OutputEffectInstance.h
#	Engine/ParallelRenderArgs.cpp
#	Engine/ParallelRenderArgs.h
#	Engine/TextureRect.cpp
#	Engine/TextureRect.h
#	Engine/UpdateViewerParams.h
#	Engine/ViewerInstancePrivate.h
#	Serialization/ImageKeySerialization.cpp
#	Serialization/ImageKeySerialization.h
#	Serialization/ImageParamsSerialization.cpp
#	Serialization/ImageParamsSerialization.h

+++ tools/genStaticDocs.sh
@@ -1,7 +1,13 @@
-#!/bin/sh
-set -x
+#!/bin/bash
+# regenerate static documentation.
+# Should be used using a snapshot build, eg:
+# env FONTCONFIG_FILE=/Applications/Natron.app/Contents/Resources/etc/fonts/fonts.conf OCIO=/Applications/Natron.app/Contents/Resources/OpenColorIO-Configs/nuke-default/config.ocio OFX_PLUGIN_PATH=/Applications/Natron.app/Contents/Plugins ./genStaticDocs.sh ~/Development/Natron-2.1/Renderer/build/Debug/NatronRenderer  /var/tmp/natrondocs ~/Development/Natron-2.1/Documentation
 
-PANDOC="$HOME/.cabal/bin/pandoc"
+#set -v
+#set -x
+set -u
+set -e
+
 NATRON_BIN=${1:-}
 TMP_FOLDER=${2:-}
 DOC_FOLDER=${3:-}
@@ -11,10 +17,17 @@ if [ -z "$NATRON_BIN" ] || [ -z "$TMP_FOLDER" ] || [ -z "$DOC_FOLDER" ]; then
   exit 1
 fi
 
+#PANDOC="$HOME/.cabal/bin/pandoc"
+PANDOC=pandoc
+SED=sed
+if [ "$(uname -s)" = "Darwin" ]; then
+    SED=gsed
+fi
+
 if [ ! -d "$TMP_FOLDER" ]; then
   mkdir -p "$TMP_FOLDER" || exit 1
 fi
-rm -rf "$TMP_FOLDER"/* || true
+rm -rf "${TMP_FOLDER:?}"/* || true
 if [ ! -d "$TMP_FOLDER/plugins" ]; then
   mkdir -p "$TMP_FOLDER/plugins" || exit 1
 fi
@@ -25,32 +38,34 @@ if [ ! -d "$DOC_FOLDER/source/plugins" ]; then
   mkdir -p "$DOC_FOLDER/source/plugins" || exit 1
 fi
 
-touch dummy.ntp
-"$NATRON_BIN" --export-docs "$TMP_FOLDER" dummy.ntp
-cd "$TMP_FOLDER" || exit 1
+touch "$TMP_FOLDER/dummy.ntp"
+"$NATRON_BIN" --export-docs "$TMP_FOLDER" "$TMP_FOLDER/dummy.ntp"
+rm "$TMP_FOLDER/dummy.ntp"
 
-for i in *.md;do $PANDOC $i --columns=1000 -o `echo $i|sed 's/.md/.rst/'`;done
-for x in plugins/*.md;do
-  $PANDOC $x --columns=9000 -o `echo $x|sed 's/.md/.rst/'`
-  PLUG=`echo $x|sed 's/.md//;s#plugins/##'`
-  sed -i "1i.. _${PLUG}:\n" plugins/${PLUG}.rst
+pushd "$TMP_FOLDER"
+for i in *.md; do
+    $PANDOC "$i" --columns=1000 -o "$(echo "$i"|$SED 's/\.md$/.rst/')"
 done
+for x in plugins/*.md; do
+    $PANDOC "$x" --columns=9000 -o "$(echo "$x"|$SED 's/\.md$/.rst/')"
+    PLUG="$(echo "$x"|$SED 's/.md//;s#plugins/##')"
+    $SED -i "1i.. _${PLUG}:\n" plugins/"${PLUG}".rst
+done
+popd
 
-for y in *.rst; do
-    cp "$TMP_FOLDER"/$y "$DOC_FOLDER/source/" || exit 1
+for y in "$TMP_FOLDER"/*.rst; do
+    cp "$y" "$DOC_FOLDER/source" || exit 1
 done
-for z in plugins/*.rst; do
+for z in "$TMP_FOLDER"/plugins/*.rst; do
   # Replace custom link, example: [linux](|html::/guide/linux.html#linux||rst::linux<Linux>|)
   # removes html::* and convert rst::* to a proper RST link, example: :ref:`linux<Linux>`
-  sed -i 's/<|html::.*||/|/g;s/|rst::\(.*\)|>.*__/:ref:\`\1\`/g' ${z}
+  $SED -i 's/<|html::.*||/|/g;s/|rst::\(.*\)|>.*__/:ref:\`\1\`/g' "${z}"
 
-  cp "$TMP_FOLDER"/$z "$DOC_FOLDER/source/plugins/" || exit 1
+  cp "$z" "$DOC_FOLDER/source/plugins" || exit 1
 done
 
-for x in plugins/*.png; do 
-cp "$TMP_FOLDER"/$x "$DOC_FOLDER/source/plugins/" || exit 1
+for x in "$TMP_FOLDER"/plugins/*.png; do 
+cp "$x" "$DOC_FOLDER/source/plugins" || exit 1
 done
 
-cd  .. || exit 1
-
-rm -rf $TMP_FOLDER dummy.ntp
+rm -rf "$TMP_FOLDER"
